docker_builder.py - update

+++ docker/docker_builder.py
@@ -14,17 +14,17 @@ def copy_file(src, dst, file_name):
 
 
 def rm_dir(dst, dir_name):
-    try:
-        rmtree(str(dst / dir_name))
-    except Exception as e:
-        pass
+    d = dst / dir_name
+    if not os.path.isdir(d):
+        return
+    rmtree(str(dst / dir_name))
 
 
 def rm_file(dst, file_name):
-    try:
-        os.remove(str(dst / file_name))
-    except Exception as e:
-        pass
+    file = dst / file_name
+    if not os.path.isfile(file):
+        return
+    os.remove(str(file))
 
 
 def copy_to_backend(parent_dir):
